+++ .changeset/three-aliens-return.md
@@ -1,0 +1,9 @@
+---
+'@yukinu/validators': patch
+'@yukinu/dashboard': patch
+'@yukinu/api': patch
+'@yukinu/ui': patch
+'@yukinu/web': patch
+---
+
+added reciate product variants
