+++ apps/dashboard/src/routes/_index.tsx
@@ -1,13 +1,7 @@
 export default function HomePage() {
   return (
     <main className='container flex-1 py-4'>
-      lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod
-      tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-      veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-      commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-      velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-      cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
-      est laborum.
+      <h1 className='sr-only'>Dashboard page</h1>
     </main>
   )
 }
