+++ apps/dashboard/src/routes/admin/categories/_index.tsx
@@ -40,7 +40,7 @@ export default function CategoriesPage() {
         </Button>
       </div>
 
-      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm'>
+      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm dark:border'>
         <Table>
           <TableHeader>
             <TableRow>
