+++ apps/dashboard/src/routes/admin/users/_index.tsx
@@ -33,7 +33,7 @@ export default function UsersPage() {
         </Button>
       </div>
 
-      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm'>
+      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm dark:border'>
         <Table>
           <TableHeader>
             <TableRow>
